+++ personalwiki/js/api.js
@@ -1,5 +1,6 @@
 
 import {v4 as uuid } from 'uuid';
+import sha1 from 'sha1';
 
 /*
 	Define functions to fetch the cells, files 
@@ -10,32 +11,24 @@ import {v4 as uuid } from 'uuid';
 					 computation.												*/
 			
 
-async function fetch_cells(notebook, file) {
-	/*
-	fetch('/api/cells/').then(res => res.json()).then( (result) => {
-		return filter_cells(file, result);
-	}, (err) => {
-		handle_cell_fetch_failure(error);
+function fetch_cells(notebook, file) {
+	return new Promise(function(resolve, reject) {
+		var req = new XMLHttpRequest();
+		var url = 'http://localhost:8000/api/file/' + WORKING_FILE + '/cells';
+
+		req.open('POST', url, true);
+		req.setRequestHeader('Content-Type', 'application/json');
+		req.onreadystatechange = () => {
+			if( req.readyState === 4 ) {
+				var json = JSON.parse(req.responseText);
+				resolve(json);
+			}
+		};
+
+		req.send();
 	});
-	*/
-
-	// TODO: ERR func
-
-	const res = await fetch('/api/cells/');
-	const json = await res.json();
-	//return filter_cells(file, json);
-	return json;
 }
 
-function filter_cells(file, cells) {
-	var data = [];
-	for( let i=0; i < Object.keys(cells).length; i++ ) {
-		if(cells[i].main_file != null && cells[i].main_file.name == file) {
-			data.push(cells[i]);
-		}
-	}
-	return data;
-}
 
 function handle_cell_fetch_failure(error) {
 	console.log("OOPSIE, there was a problem >.<");
@@ -45,21 +38,23 @@ async function post_cell_update(uuid, data) {
 	var xhr = new XMLHttpRequest();
 	xhr.open("PUT", "http://localhost:8000/api/cell/" + uuid , true);
 	xhr.setRequestHeader("Content-Type", 'application/json');
-	xhr.send(JSON.stringify(data));
+	var payload = { "data": data, "parent_url": WORKING_FILE, "hash":sha1(data) };
+	xhr.send(JSON.stringify(payload));
 }
 
 async function delete_cell(uuid, data) {
 	var xhr = new XMLHttpRequest();
 	xhr.open("DELETE", "http://localhost:8000/api/cell/" + uuid , true);
 	xhr.setRequestHeader("Content-Type", 'application/json');
-	xhr.send(JSON.stringify(data));
+	xhr.send()
 }
 
 async function patch_cell(uuid, data) {
 	var xhr = new XMLHttpRequest();
 	xhr.open("PATCH", "http://localhost:8000/api/cell/" + uuid , true);
 	xhr.setRequestHeader("Content-Type", 'application/json');
-	xhr.send(JSON.stringify(data));
+	var payload = { "data": data, "parent_url": WORKING_FILE, "hash":sha1(data) };
+	xhr.send(JSON.stringify(payload));
 }
 
 export { fetch_cells, post_cell_update, delete_cell, patch_cell };