+++ personalwiki/wikiapp/models.py
@@ -4,13 +4,9 @@ from django.utils.translation import ugettext_lazy as _
 from taggit.managers import TaggableManager
 from taggit.models import GenericUUIDTaggedItemBase, TaggedItemBase
 
-
 import uuid
+from datetime import datetime
 
-""" I am not sure if the cells should be stored separately
-    or simply in a single data field. Max theoretical file
-    size could get pretty big.
-    Each cell is part of a File, and each File is part of a Notebook """
 class Notebook(models.Model):
     title = models.CharField(max_length=120)
 
@@ -19,12 +15,12 @@ class Notebook(models.Model):
 
 class File(models.Model):
     name = models.CharField(max_length=200)
-    #last_edit = models.DateTimeField(auto_now_add=True, null=True)
+    last_edit = models.DateTimeField(auto_now=True, null=True)
+    url = models.SlugField(max_length=200, unique=True)
     notebook = models.ForeignKey(Notebook, on_delete=models.SET_NULL, null=True, blank=True)
 
     def __str__(self):
         return self.name
-
 
 """ The cell model has both a uuid and a hash :
     The uuid is for identifying if the cell is in the db.
@@ -34,7 +30,7 @@ class Cell(models.Model):
     data = models.TextField()
     uuid = models.UUIDField(primary_key=True, unique=True, default=uuid.uuid4, editable=False)
     uhash = models.CharField(max_length=120, default="NULL")
-    #main_file = models.ForeignKey(File, on_delete=models.SET_NULL, null=True, blank=True)
+    mf = models.ForeignKey(File, on_delete=models.SET_NULL, null=True)
 
     def __str__(self):
         return self.data