+++ personalwiki/wikiapp/serializers.py
@@ -2,6 +2,7 @@
 
 from rest_framework import serializers
 from wikiapp.models import Cell, File, Notebook
+from django.utils.text import slugify
 
 # Serialize the data before sending it between the database and UI
 
@@ -9,7 +10,7 @@ class NotebookSerializer(serializers.Serializer):
     title = serializers.CharField(max_length=120)
 
     def create(self, validated_data):
-        return Notebook.objects.create(**validated_data)
+        instance = Notebook.objects.create(**validated_data)
 
     def update(self, instance, validated_data):
         instance.title = validated_data.get('title', instance.title)
@@ -18,14 +19,18 @@ class NotebookSerializer(serializers.Serializer):
 
     class Meta:
         model = Notebook
+        fields = ('title')
 
 
 class FileSerializer(serializers.Serializer):
     name = serializers.CharField(max_length=200)
-    #last_edit = serializers.DateTimeField()
+    last_edit = serializers.DateTimeField()
+    url = serializers.CharField(max_length=200)
     notebook = NotebookSerializer()
 
     def create(self, validated_data):
+        pn = Notebook.objects.filter(title=validated_data['notebook']['title'])
+        validated_data['notebook'] = pn[0];
         return File.objects.create(**validated_data)
 
     def update(self, instance, validated_data):
@@ -39,19 +44,20 @@ class FileSerializer(serializers.Serializer):
         model = File
 
 
-    #main_file = FileSerializer()
 class CellSerializer(serializers.Serializer):
     data = serializers.CharField(style={'base_template' : 'textarea.html'})
     uuid = serializers.CharField()
     uhash = serializers.CharField()
+    mf = FileSerializer()
 
     def create(self, validated_data):
+        pf = File.objects.filter(url=validated_data['mf']['url'])
+        validated_data['mf'] = pf[0]
         return Cell.objects.create(**validated_data)
 
     def update(self, instance, validated_data):
         instance.data = validated_data.get('data', instance.data)
         instance.uuid = validated_data.get('uuid', instance.uuid)
-        #instance.main_file = validated_data.get('main_file', instance.main_file)
         instance.save()
         return instance