+++ start.sh
@@ -1,6 +1,6 @@
 
 if python -c "import django, rest_framework" &> /dev/null; then
-	echo 'Requirements already installed.'
+	echo ' Python requirements already satisfied.'
 else
 	pip install -r ./requirements.txt
 fi