+++ changelog.md
@@ -1,10 +1,30 @@
 
 ### Changelog
 
-#### 0.0.1
+&nbsp;
+
+#### 0.2.0
+
+New Features:
+* Multiple File Support
+* Easily Customisable Themes
+
+Improvements:
+* Better API endpoints
+* API Testing
+* Default directory tree page
+* Improved Markdown syntax
+
+Style Changes:
+* Removed Page Borders
+* Added selection and editing color coding border.
+* Added Cell indexing
+
+&nbsp;
+
+#### 0.1.0
 * Cell rendering supports Markdown and LaTeX using `react-markdown` and `react-mathjax`.
 * Cell selection can MOVE and DELETE individual cells.
 * Using Django Backend with the prepackaged WSGI Python webserver.
 * Storing data in a SQLite database.
 * Using `djangorestframework` to recieve and serve data.
-*